Add files via upload
Ram Mandir Chronicles 🚀

Overview:
The "Ram Mandir Chronicles" is a digital journey that encapsulates the 500-year saga of the Ram Mandir in Ayodhya. This project aims to provide a visual and narrative exploration of the historical, cultural, and architectural milestones associated with the temple's journey.

Features:
Historical Timeline: Dive into a digital timeline that unfolds the key events over the past five centuries.
Interactive Visuals: Experience the story through a combination of historical images, maps, and architectural renderings.
Educational Purpose: This project is developed for educational purposes, offering a learning experience about the rich history of the region.

Credits:
Satyanshu Bhardwaj
CodePen, ALjaezeera, W3Sschools,ChatGPT
Disclaimer:
This project is created solely for learning purposes and cultural exploration. It does not intend to hurt any sentiments or promote any particular viewpoint. Respect for diverse perspectives and histories is encouraged.

Feel free to contribute, share insights, and explore the rich cultural heritage embedded in the Ram Mandir Chronicles!